requirements.txt
zhconv
requests
PyGithub